requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
Django==3.1.7
django-appconf==1.0.4
django-countries==7.0
django-extensions==3.1.1
django-filter==2.4.0
django-imagekit==4.0.2
idna==2.10
kavenegar==1.1.2
phonenumbers==8.12.20
pilkit==2.0
Pillow==8.1.0
pytz==2021.1
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.4
